+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 
-- * 2d31da5 - 2018-04-29: Release v9.3.10 
+- * bdf89e6 - 2018-04-29: Release v9.3.11 
+- * ae29eb5 - 2018-04-29: Fixed log level of sentry logging handler to log errors only 
+- * bebb5f6 - 2018-04-29: Pre-Release v9.3.11.dev1 
+- * a82f305 - 2018-04-29: Fixed issue with NewPCT provider and &#x27;NoneType&#x27; object has no attribute &#x27;get&#x27; Fixed missing method from twilio client 
+- * bcac79f - 2018-04-29: Release v9.3.10 
 - * 79bf6f6 - 2018-04-29: Release v9.3.8 
 - * fd43837 - 2018-04-29: Release v9.3.7 
 - * b9546e2 - 2018-04-29: Pre-Release v9.3.7.dev3